requirements.txt
decord; platform_machine != 'arm64'
eva-decord; platform_machine == 'arm64'
gradio
huggingface_hub
imageio
matplotlib
numpy==1.26.4
omegaconf
openai
opencv-python>=4.4.0.46
openpyxl
pandas
pillow
portalocker
protobuf
python-dotenv
requests
rich
sentencepiece
setuptools
sty
tabulate
tiktoken
timeout-decorator
torch
torchvision
tqdm
transformers
typing_extensions
validators
xlsxwriter
ftfy
regex
pandas
decord
scikit-learn
hydra-core
google-generativeai
flash-attn==2.6.3
qwen-vl-utils
open_clip_torch
clip